created PersonSearchResults POM and moved relevant methods across
