+++ PersonSearchDetails/WCCIS.Specs.csproj
@@ -98,6 +98,7 @@
     <Compile Include="Enums\Enums.cs" />
     <Compile Include="Extensions\DHCWExtensions.cs" />
     <Compile Include="Hooks\Hooks1.cs" />
+    <Compile Include="PageObjects\Page_PersonSearchResults.cs" />
     <Compile Include="PageObjects\Page_CommClinicReferral.cs" />
     <Compile Include="PageObjects\Page_MPISearch.cs" />
     <Compile Include="PageObjects\Page_PersonCoreDemographics.cs" />
